+++ react/components/SKUSelector/Wrapper.tsx
@@ -162,6 +162,7 @@ interface Props {
   thumbnailImage?: string
   visibleVariations?: string[]
   showVariationsLabels?: ShowVariationsLabels
+  variationsWithSingleValueAreVisibles?: boolean
   variationsSpacing?: number
   showVariationsErrorMessage?: boolean
   initialSelection?: InitialSelectionType
@@ -254,6 +255,7 @@ function SKUSelectorWrapper(props: Props) {
         variationsSpacing={props.variationsSpacing}
         showValueForVariation={showValueForVariation}
         showVariationsLabels={props.showVariationsLabels}
+        variationsWithSingleValueAreVisibles={props.variationsWithSingleValueAreVisibles}
         hideImpossibleCombinations={props.hideImpossibleCombinations}
         disableUnavailableSelectOptions={props.disableUnavailableSelectOptions}
         showVariationsErrorMessage={props.showVariationsErrorMessage}